fix clippy

+++ src/telegram/actions/global_stats.rs
@@ -16,10 +16,10 @@ pub async fn handle(
     m: &Message,
 ) -> anyhow::Result<HandleStatus> {
     let disliked =
-        TrackStatusService::count_status(app.db(), TrackStatus::Disliked, None, None).await? as u64;
+        TrackStatusService::count_status(app.db(), TrackStatus::Disliked, None, None).await?;
     let ignored =
-        TrackStatusService::count_status(app.db(), TrackStatus::Ignore, None, None).await? as u64;
-    let skipped = TrackStatusService::sum_skips(app.db(), None).await? as u64;
+        TrackStatusService::count_status(app.db(), TrackStatus::Ignore, None, None).await?;
+    let skipped = TrackStatusService::sum_skips(app.db(), None).await?;
 
     let UserStats {
         removed_collection,
